Rename files to remove CA convention (#4714)
